SMA refractor(Time complexity:O(1) for panning or live data flow)

+++ include/candle.h
@@ -21,8 +21,7 @@ public:
     
 };
 
-int get_highest_price(vector<candle> &data,Viewport &Viewport);
-int get_lowest_price(vector<candle> &data,Viewport &Viewport);
-
-int get_highest_volume(vector<candle> &data,Viewport &Viewport);
-int get_lowest_volume(vector<candle> &data,Viewport &Viewport);
+double get_highest_price(vector<candle> &data,Viewport &Viewport);
+double get_lowest_price(vector<candle> &data,Viewport &Viewport);
+double get_highest_volume(vector<candle> &data,Viewport &Viewport);
+double get_lowest_volume(vector<candle> &data,Viewport &Viewport);
