+++ src/candle.cpp
@@ -1,8 +1,8 @@
 #include "../include/candle.h"
 
-int get_highest_price(vector<candle> &data,Viewport &Viewport){
-    int highest_price=0;
-    int end = min((Viewport.first_visible_candle + Viewport.candle_count),int(data.size()));
+double get_highest_price(vector<candle> &data,Viewport &Viewport){
+    double highest_price=0;
+    double end = min((Viewport.first_visible_candle + Viewport.candle_count),int(data.size()));
 
     for(int i=Viewport.first_visible_candle;i<end;i++){
         if(data[i].high_price>highest_price){
@@ -13,10 +13,10 @@ int get_highest_price(vector<candle> &data,Viewport &Viewport){
     return highest_price;
 }
 
-int get_lowest_price(vector<candle> &data,Viewport &Viewport){
+double get_lowest_price(vector<candle> &data,Viewport &Viewport){
     
-    int lowest_price = data[0].low_price;
-    int end = min((Viewport.first_visible_candle + Viewport.candle_count),int(data.size()));
+    double lowest_price = data[0].low_price;
+    double end = min((Viewport.first_visible_candle + Viewport.candle_count),int(data.size()));
 
     for(int i=Viewport.first_visible_candle;i<end;i++){
         if(data[i].low_price<lowest_price){
@@ -27,10 +27,10 @@ int get_lowest_price(vector<candle> &data,Viewport &Viewport){
     return lowest_price;
 }
 
-int get_highest_volume(vector<candle> &data,Viewport &Viewport){
+double get_highest_volume(vector<candle> &data,Viewport &Viewport){
 
-    int highest_volume=data[Viewport.first_visible_candle].volume;
-    int end = min((Viewport.first_visible_candle + Viewport.candle_count),int(data.size()));
+    double highest_volume=data[Viewport.first_visible_candle].volume;
+    double end = min((Viewport.first_visible_candle + Viewport.candle_count),int(data.size()));
 
     for(int i=Viewport.first_visible_candle+1;i<end;i++){
         if(data[i].volume>highest_volume){
@@ -43,10 +43,10 @@ int get_highest_volume(vector<candle> &data,Viewport &Viewport){
 
 }
 
-int get_lowest_volume(vector<candle> &data,Viewport &Viewport){
+double get_lowest_volume(vector<candle> &data,Viewport &Viewport){
     
-    int lowest_volume = data[Viewport.first_visible_candle].volume;
-    int end = min((Viewport.first_visible_candle + Viewport.candle_count),int(data.size()));
+    double lowest_volume = data[Viewport.first_visible_candle].volume;
+    double end = min((Viewport.first_visible_candle + Viewport.candle_count),int(data.size()));
 
     for(int i=Viewport.first_visible_candle+1;i<end;i++){
         if(data[i].volume<lowest_volume){
