+++ src/grid.cpp
@@ -1,14 +1,14 @@
 #include "../include/grid.h"
 
-int scale(GridConfig &config,int value,int highest,int lowest){
+int scale(GridConfig &config,double value,double highest,double lowest){
 
-    int gap=highest-lowest;
+    double gap=highest-lowest;
 
     if(gap==0){
         return config.chart_height/2;
     }
     else{
-    int scaled_val=(value-lowest)*(config.chart_height-1)/gap;
+    double scaled_val=(value-lowest)*(config.chart_height-1)/gap;
 
     return config.chart_height - 1 - scaled_val;
     }
